lyx: update to 2.3.2
Second Maintenance Release in 2.3.x Series: LyX 2.3.2 released.
December 14, 2018

LyX 2.3.2 is the result of on-going efforts to make our stable version more
reliable and more stable. One main fix involves a crash with external viewers
on Windows. We have also fixed a number of problems with the output of right-
to-left languages, thanks to the involvement of users. There was also an
infinite loop bug connected with the separation of environments that has been
fixed. But, for the most part, the fixes here are incremental.

First Maintenance Release in 2.3.x Series: LyX 2.3.1 released.
September 16, 2018

LyX 2.3.1 is the result of on-going efforts to make our stable version more
reliable and more stable. We have fixed a number of bugs and added some new
features. Perhaps the most important of these is that Jean-Marc Lasgouttes
re-wrote the document painting mechanism.
This makes LyX snappier, especially on repeated events.

All python scripts distributed with LyX should now be compatible with both
python 2.x and python 3.x.

New Stable Series: LyX 2.3.0 released.
March 16, 2018

LyX 2.3.0 is the culmination of two years of hard work. An overview
of the new features can be found here:
  https://wiki.lyx.org/LyX/NewInLyX23